upd: add caching for reply fetch, refactor tests
