+++ fast_mda_traceroute/formats/traceroute.py
@@ -18,7 +18,10 @@ def format_traceroute(replies: List[Reply]) -> str:
         replies, lambda x: x.probe_ttl, lambda x: x.reply_src_addr, set
     )
     output = ""
-    for ttl in range(min(nodes_by_ttl), min(max(nodes_by_ttl), destination_ttl) + 1):
+    for ttl in range(
+        min(nodes_by_ttl, default=0),
+        min(max(nodes_by_ttl, default=255), destination_ttl) + 1,
+    ):
         output += f"{ttl} "
         output += " ".join(nodes_by_ttl.get(ttl, []))
         output += "\n"
