+++ tests/fakenet/fakenet.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 from dataclasses import dataclass
-from typing import Optional
 import xxhash
 import random
 
@@ -267,7 +266,7 @@ class FakeProber:
         return replies
 
 
-@dataclass
+@dataclass(frozen=True)
 class FakeReply:
     """
     A fake reply that emulates the replies generated by the prober.
@@ -339,36 +338,19 @@ class FakeReply:
         :return: a FakeReply instance
         """
         # return a fake ICMP time ECHO reply using information from the probe
-        # re-uses the icmp_time_exceeded_reply method
-        reply = FakeReply.icmp_time_exceeded_reply(src, probe)
-        reply.reply_icmp_type = 0
-        return reply
-
-
-def graph_from_links(
-    src_addr: str, links: dict[int, list[tuple[str, str]]]
-) -> nx.DiGraph:
-    """
-    Create a networkx graph from a dictionary of links.
-    The dictionary has the following structure:
-    {
-        ttl: [(src, dst), ...],
-        ...
-    }
-    Since the links dictionnary starts at TTL 1,
-    we need to connect the source address to all nodes at TTL 1.
-    :param src_addr: the source address of the network
-    :param links: a dictionary of links by TTL
-
-    :return: a networkx DiGraph instance
-    """
-    nx_graph = nx.DiGraph()
-    # link source address to all nodes at ttl 1
-    for link in links[1]:
-        nx_graph.add_edge(src_addr, link[0])
-
-    # add all edges to the graph
-    for _, links in links.items():
-        nx_graph.add_edges_from(list(set(links)))
-
-    return nx_graph
+        # does not reuse because FakeReply is immutable (frozen)
+        return FakeReply(
+            probe_protocol=probe.protocol,
+            probe_dst_addr=probe.dst_addr,
+            probe_src_port=probe.src_port,
+            probe_dst_port=probe.dst_port,
+            probe_ttl=probe.ttl,
+            reply_ttl=255 - probe.ttl,
+            reply_src_addr=src,
+            reply_icmp_type=0,
+            reply_icmp_code=0,
+            reply_id=FakeProber.flow_id(probe, 0) % 2**16,
+            quoted_ttl=1,
+            capture_timestamp=0,
+            rtt=0,
+        )
