+++ tests/test_links.py
@@ -17,7 +17,7 @@ def test_get_replies_by_flow(make_reply):
         make_reply(1, "::9", 24001, 33434, 1, "::1"),
         make_reply(1, "::9", 24001, 33434, 3, "::3"),
     ]
-    assert get_replies_by_flow([*replies_flow_1, *replies_flow_2]) == {
+    assert get_replies_by_flow((*replies_flow_1, *replies_flow_2)) == {
         (1, "::9", 24000, 33434): replies_flow_1,
         (1, "::9", 24001, 33434): replies_flow_2,
     }
@@ -35,7 +35,7 @@ def test_get_replies_by_ttl(make_reply):
         make_reply(1, "::9", 24000, 33434, 3, "::3"),
         make_reply(1, "::9", 24001, 33434, 3, "::3"),
     ]
-    assert get_replies_by_ttl([*replies_ttl_1, *replies_ttl_2, *replies_ttl_3]) == {
+    assert get_replies_by_ttl((*replies_ttl_1, *replies_ttl_2, *replies_ttl_3)) == {
         1: replies_ttl_1,
         2: replies_ttl_2,
         3: replies_ttl_3,
@@ -52,7 +52,7 @@ def test_get_pairs_by_flow(make_reply):
         make_reply(1, "::9", 24001, 33434, 1, "::1"),
         make_reply(1, "::9", 24001, 33434, 3, "::3"),
     ]
-    assert get_pairs_by_flow([*replies_flow_1, *replies_flow_2]) == {
+    assert get_pairs_by_flow((*replies_flow_1, *replies_flow_2)) == {
         (1, "::9", 24000, 33434): [
             (1, replies_flow_1[0], replies_flow_1[1]),
             (2, replies_flow_1[1], replies_flow_1[2]),
@@ -74,7 +74,7 @@ def test_get_links_by_ttl(make_reply):
         make_reply(1, "::9", 24001, 33434, 1, "::1"),
         make_reply(1, "::9", 24001, 33434, 3, "::3"),
     ]
-    links_by_ttl = get_links_by_ttl([*replies_flow_1, *replies_flow_2])
+    links_by_ttl = get_links_by_ttl((*replies_flow_1, *replies_flow_2))
     links_by_ttl = {k: set(v) for k, v in links_by_ttl.items()}
     assert links_by_ttl == {
         1: {(1, "::1", "::2"), (1, "::1", None)},
@@ -92,7 +92,7 @@ def test_get_scamper_links(make_reply):
         make_reply(1, "::9", 24001, 33434, 1, "::1"),
         make_reply(1, "::9", 24001, 33434, 3, "::3"),
     ]
-    assert get_scamper_links([*replies_flow_1, *replies_flow_2]) == {
+    assert get_scamper_links((*replies_flow_1, *replies_flow_2)) == {
         "::1": {
             2: {"::2": [replies_flow_1[1]], None: [None]},
             3: {"::3": [replies_flow_2[1]]},
